pyproject.toml
[project]
name = "capricorn-adk-agent"
version = "0.1.0"
description = "Medical oncology agent for genetic and oncological patient treatment recommendations"
authors = [
        { name = "Capricorn Medical Team", email = "medical@capricorn.com" },
]

license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic-settings = "^2.8.1"
tabulate = "^0.9.0"
google-auth = "^2.36.0"
requests = "^2.32.3"
google-adk = "^1.0.0"
google-cloud-aiplatform = { extras = [
        "adk",
        "agent-engines",
], version = "^1.93.0" }

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
pytest-cov = "^6.0.0"
pytest-asyncio = "^0.26.0"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"